refactor: move code in core.py to separate files
